requirements.txt
boto3==1.26.9
botocore==1.29.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
paho-mqtt==1.6.1
pandas
PyMySQL==1.0.2
psycopg2-binary==2.9.9
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.1
